move update check from settings to admin
